Refactor error_report functions to static methods
